NOISSUE: move integr tests from service network
