+++ network/tests/suite_test.go
@@ -50,7 +50,7 @@
 
 // +build networktest
 
-package servicenetwork
+package tests
 
 import (
 	"context"
@@ -64,9 +64,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/ThreeDotsLabs/watermill/message"
+	"github.com/insolar/insolar/network/servicenetwork"
 	"github.com/stretchr/testify/suite"
-
-	"github.com/insolar/insolar/network/gateway"
 
 	"github.com/insolar/insolar/certificate"
 	"github.com/insolar/insolar/component"
@@ -322,7 +322,7 @@ type networkNode struct {
 	ctx                 context.Context
 
 	componentManager *component.Manager
-	serviceNetwork   *ServiceNetwork
+	serviceNetwork   *servicenetwork.ServiceNetwork
 	consensusResult  chan error
 }
 
@@ -438,6 +438,16 @@ func (m staterMock) State() ([]byte, error) {
 	return m.stateFunc()
 }
 
+type PublisherMock struct{}
+
+func (p *PublisherMock) Publish(topic string, messages ...*message.Message) error {
+	return nil
+}
+
+func (p *PublisherMock) Close() error {
+	return nil
+}
+
 // preInitNode inits previously created node with mocks and external dependencies
 func (s *testSuite) preInitNode(node *networkNode) {
 	t := s.T()
@@ -453,10 +463,10 @@ func (s *testSuite) preInitNode(node *networkNode) {
 
 	node.componentManager = &component.Manager{}
 	node.componentManager.Register(platformpolicy.NewPlatformCryptographyScheme())
-	serviceNetwork, err := NewServiceNetwork(cfg, node.componentManager, false)
+	serviceNetwork, err := servicenetwork.NewServiceNetwork(cfg, node.componentManager, false)
 	s.Require().NoError(err)
 
-	serviceNetwork.SetGateway(gateway.NewFakeOk())
+	serviceNetwork.SetGateway(NewFakeOk())
 	serviceNetwork.Gateway().Run(context.Background())
 
 	amMock := staterMock{
